+++ lib/capistrano-abril/bundler.rb
@@ -8,6 +8,7 @@ CapistranoAbril.with_configuration do
     task :install do
 
       run "mkdir -p #{shared_path}/bundle"
+<<<<<<< HEAD
 #
 
 
@@ -18,6 +19,17 @@ CapistranoAbril.with_configuration do
       else
           # stage/prod
           run "bundle install --gemfile #{latest_release}/Gemfile --path #{shared_path}/bundle --without development test"
+=======
+      
+      env = fetch(:rails_env, fetch(:env, 'production'))
+      
+      if ['production', "stage"].include?(env)
+        # stage/prod
+        run "bundle install --gemfile #{latest_release}/Gemfile --path #{shared_path}/bundle --deployment --without development test"
+      else
+        # dev/qa
+        run "bundle install --gemfile #{latest_release}/Gemfile --path #{shared_path}/bundle --deployment"
+>>>>>>> d5601ae1eb19935018b29a0edea896b1d123c881
       end
 
     end
@@ -26,4 +38,3 @@ CapistranoAbril.with_configuration do
   end # namespace
 
 end
-