+++ lib/capistrano-abril/check_rails3.rb
@@ -1,20 +1,8 @@
 require File.dirname(__FILE__) + '/../capistrano-abril' if ! defined?(CapistranoAbril)
+require File.dirname(__FILE__) + '/check'
 
 CapistranoAbril.with_configuration do
 
-  namespace :check do
-
-    desc '(check-rails3.rb) Check Rails3 before symlink/startup.'
-    task :rails3, :roles => :app, :except => { :no_release => true } do
-      on_rollback { run "rm -rf #{latest_release}; true" }
-
-      run "cd #{latest_release} && RAILS_ENV=#{rails_env} bundle exec rails runner 'puts Rails.env'"
-
-    end
-    before "deploy:symlink", "check:rails3"
-
-  end # namespace
+  before "deploy:symlink", "check:rails3"
 
 end
-
-